step3_db_beckend_gunicorn/library/requirements.txt
Django==2.2.24
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.11.2
djangorestframework-simplejwt
drf-yasg==1.20
Markdown
graphene
graphene-django
psycopg2-binary
mixer


